Recognize storyboards as XML files
This is to allow customization of the CDVLaunchScreen.storyboard to
adjust contentmode and background color in config.xml.

Further:

I'm not quite sure why the logic is that only some extensions are
recognized as XML and all others as plists. Maybe the opposite would be
better?

Alternatively, an additional parameter to <config-file> and
<edit-config> to specifiy the file format could be useful.