use fast-glob for file globbing
Until now we used our own glob implementation based on minimatch.

The only difference to the old implementation is, that not every parent
directory of a matched file is included in the result. However, since do
not pass the dirMaps out of this module and any missing dirs will be
created by fs-extra, that should be an advantage, if anything.