fix(ConfigParser): ImageResources constructor (#142)
The previous ImageResources constructor did not cover the case where a
single integer argument is given to construct an empty array of the
given length. This broke the `map` method since it uses that constructor
variant.

This commit fixes that situation by using a static factory method
instead of overriding the native Array constructor.