refactor(bom): Bring strip-bom in as util
The strip-bom npm package moved to mandatory ES Modules, which means
we're stuck on an old version and it shows up every time we run `npm
outdated` as something to be checked. The actual code for stripping the
BOM is a handful of lines of code that a stable and well-solved problem,
so it should be fine to just bring this one function into cordova-common
as a util rather than depending on a 3rd party module to provide it.

In a better world, there would be some built-in NodeJS option to handle
this...